flowedit: Extract undo_redo module from main.rs
Move record_edit, apply_undo, and apply_redo functions into a dedicated
undo_redo.rs module. These functions handle edit history recording and
replay for undo/redo operations.
